+++ src/patch/core/SharedDef.zig
@@ -71,7 +71,7 @@ pub const GlobalState = extern struct {
     } = .{},
 };
 
-pub const GLOBAL_FUNCTION_VERSION = 18;
+pub const GLOBAL_FUNCTION_VERSION = 19;
 
 pub const GlobalFunction = extern struct {
     // Settings
@@ -90,12 +90,12 @@ pub const GlobalFunction = extern struct {
     InputGetXInputAxis: *const fn (axis: XINPUT_GAMEPAD_AXIS_INDEX) f32,
     // Game
     // TODO: change prefix to G (analogous to R for Resource)
-    GameFreezeEnable: *const fn (o: [*:0]const u8) bool,
-    GameFreezeDisable: *const fn (o: [*:0]const u8) bool,
-    GameFreezeIsFrozen: *const fn () bool,
-    GameHideRaceUIEnable: *const fn (o: [*:0]const u8) bool,
-    GameHideRaceUIDisable: *const fn (o: [*:0]const u8) bool,
-    GameHideRaceUIIsHidden: *const fn () bool,
+    GFreezeEnable: *const fn (o: [*:0]const u8) bool,
+    GFreezeDisable: *const fn (o: [*:0]const u8) bool,
+    GFreezeIsFrozen: *const fn () bool,
+    GHideRaceUIEnable: *const fn (o: [*:0]const u8) bool,
+    GHideRaceUIDisable: *const fn (o: [*:0]const u8) bool,
+    GHideRaceUIIsHidden: *const fn () bool,
     // Toast
     ToastNew: *const fn (text: [*:0]const u8, color: u32) callconv(.C) bool,
     // Resources